+++ CHANGELOG.md
@@ -1,5 +1,48 @@
 # Changelog
 
+## [1.0.3](https://github.com/KemingHe/OSU/compare/v1.0.2...v1.0.3) (2024-08-02)
+
+
+### Bug Fixes
+
+* **csv2ts.test.ts:** further removed error snapshot to accommodate Actions file path ([29dbdbb](https://github.com/KemingHe/OSU/commit/29dbdbbd3024c198781ef07b545d6d61925eca9c))
+
+## [1.0.2](https://github.com/KemingHe/OSU/compare/v1.0.1...v1.0.2) (2024-08-02)
+
+## [1.0.1](https://github.com/KemingHe/OSU/compare/v1.0.0...v1.0.1) (2024-08-02)
+
+
+### Bug Fixes
+
+* **scripts/utils/csv2ts.test.ts:** removed snapshot test to accommodate diff test env file paths ([f461757](https://github.com/KemingHe/OSU/commit/f4617575b4292c5c751830cb3ab912d75938e825))
+
+## [1.0.0](https://github.com/KemingHe/OSU/compare/v0.4.0...v1.0.0) (2024-08-02)
+
+
+### ⚠ BREAKING CHANGES
+
+* **tsup.config.ts:** to unify exports, removed `osu` from default export, added `osu` back as a name
+export, now must use `import { osu } from "@keminghe/osu";` to use the osu namespace and subsequent
+namespaces, classes, and constants.
+
+### Features
+
+* **tsup.config.ts:** adopted tsup as bundler, enforced uniform named exports ([ef24365](https://github.com/KemingHe/OSU/commit/ef2436513c390c9cb203a2f0bd5cdfbe85f34e34))
+
+## [0.4.0](https://github.com/KemingHe/OSU/compare/v0.3.0...v0.4.0) (2024-08-01)
+
+
+### Features
+
+* **csv2ts.ts:** extended generic csv 2 ts array util funtions ([7537250](https://github.com/KemingHe/OSU/commit/75372500acb1a3c268d50a34fbc3d1c92a3993b1))
+
+## [0.3.0](https://github.com/KemingHe/OSU/compare/v0.2.0...v0.3.0) (2024-08-01)
+
+
+### Features
+
+* **scripts dir:** fully implemented and tested csv to ts auto generation scripts ([679bcb4](https://github.com/KemingHe/OSU/commit/679bcb4212a04b1dc22dbfa69186881a3f9d002f))
+
 ## 0.2.0 (2024-07-31)